students/K33422/Daria Plotskaya/tourfirm/requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.0
defusedxml==0.7.1
distlib==0.3.4
Django==3.2.11
django-cors-headers==3.10.1
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
filelock==3.4.0
ghp-import==2.0.2
idna==3.3
importlib-metadata==4.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.0.1
mergedeep==1.3.4
mkdocs==1.2.3
mkdocs-material==8.1.3
mkdocs-material-extensions==1.0.3
oauthlib==3.1.1
packaging==21.3
platformdirs==2.4.0
psycopg2-binary==2.9.3
pycparser==2.21
Pygments==2.10.0
PyJWT==2.3.0
pymdown-extensions==9.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-magic==0.4.27
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
signals==0.0.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.10.0
watchdog==2.1.6
zipp==3.6.0